Refactor text_classification, improve extract_archive and add a small example (#565)

+++ torchtext/utils.py
@@ -4,6 +4,7 @@ import csv
 from tqdm import tqdm
 import os
 import tarfile
+import logging
 
 
 def reporthook(t):
@@ -42,6 +43,11 @@ def download_from_url(url, path):
     def process_response(r):
         chunk_size = 16 * 1024
         total_size = int(r.headers.get('Content-length', 0))
+        download_dir = os.path.dirname(path)
+        if not os.path.exists(download_dir):
+            raise RuntimeError(
+                "Download directorty {} does not exist. "
+                "Did you create it?".format(download_dir))
         with open(path, "wb") as file:
             with tqdm(total=total_size, unit='B',
                       unit_scale=1, desc=path.split('/')[-1]) as t:
@@ -55,7 +61,7 @@ def download_from_url(url, path):
         process_response(response)
         return
 
-    print('downloading from Google Drive; may take a few minutes')
+    logging.info('Downloading from Google Drive; may take a few minutes')
     confirm_token = None
     session = requests.Session()
     response = session.get(url, stream=True)
@@ -90,13 +96,17 @@ def utf_8_encoder(unicode_csv_data):
         yield line.encode('utf-8')
 
 
-def extract_archive(from_path, to_path=None, remove_finished=False):
-    """Extract tar.gz archives.
+def extract_archive(from_path, to_path=None, overwrite=False, archive='tar'):
+    """Extract archive.
 
     Arguments:
-        from_path: the path where the tar.gz file is.
-        to_path: the path where the extracted files are.
-        remove_finished: remove the original tar.gz file. Default: False
+        from_path: the path of the archive.
+        to_path: the root path of the extraced files (directory of from_path)
+        overwrite: overwrite existing files (False)
+        archive: the archive format to extract (tar)
+
+    Returns:
+        List of paths to extracted files even if not overwritten.
 
     Examples:
         >>> url = 'http://www.quest.dcs.shef.ac.uk/wmt16_files_mmt/validation.tar.gz'
@@ -108,11 +118,19 @@ def extract_archive(from_path, to_path=None, remove_finished=False):
     if to_path is None:
         to_path = os.path.dirname(from_path)
 
-    if from_path.endswith(".tar.gz"):
-        with tarfile.open(from_path, 'r:gz') as tar:
-            tar.extractall(path=to_path)
-    else:
-        raise ValueError("Extraction of {} not supported".format(from_path))
+    if archive != 'tar':
+        raise NotImplementedError("We currently only support tar achives.")
 
-    if remove_finished:
-        os.remove(from_path)
+    with tarfile.open(from_path, 'r:gz') as tar:
+        files = []
+        for file_ in tar:
+            if file_.isfile():
+                if os.path.exists(file_.name):
+                    if overwrite:
+                        tar.extract(file_, to_path)
+                else:
+                    tar.extract(file_, to_path)
+                files.append(os.path.join(to_path, file_.name))
+            else:
+                tar.extract(file_, to_path)
+        return files
